move server files to server folder
