--- packages/config/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./config-builder";
-export * from "./parser/config-parser";
